+++ teneto/classes/__init__.py
@@ -1,3 +1,4 @@
 #from teneto.classes.network import NewTemporalNetwork
 from .bids import TenetoBIDS
+from .network import TemporalNetwork
 #from teneto.classes.preproc import fMRIPreproc